Bugfixes to use of Poisson models for typing resistance mutations.
Ran of derivation set and found was undercalling resistance.
Debugged a single example and found a bug, where I was finding
a resistance allele and then overwriting. Fixed.

+++ src/predictor/staph/main.c
@@ -168,7 +168,7 @@ int main(int argc, char **argv)
 
   //given the error rate and other params, we can estimate expected depth of covg, and read-arrival rate
   // lambda_g = Depth/read_len _g means lambda on the true genome
-  double lambda_g_err_free = (bp_loaded/cmd_line->genome_size) / mean_read_length ; 
+  double lambda_g_err_free = ((double) bp_loaded/(double)(cmd_line->genome_size)) / (double) mean_read_length ; 
   
   int expected_depth 
     = (int) ( pow(1-err_rate, cmd_line->kmer_size)
